requirements.txt
PyYAML==5.3.1
aiohttp==3.6.2
tortoise-orm==0.16.11
asyncpg==0.20.1